Update for Python 3/Sopel 8 and convert to package
Besides putting the plugin file into the `sopel-irc/plugin` skeleton,
this required a few little tweaks:

* XKCDB supports HTTPS now, so let's use it for fetch & display URLs
* `lxml` seems not to like TLS, so fetch HTML with `requests` instead
  * This doesn't obviate the open issue to ditch `lxml`; it's just a
    separate project from making the plugin work on modern Python/Sopel
* Use a `r` (raw) string for literal `re` patterns

That's basically it, plus using `sopel.plugin` instead of `sopel.module`
as an import for plugin decorators (and changing the term "module" to
"plugin" in the readme).

Based on testing, this version can be published as-is since it works
just fine, and migrating off of `lxml` can be done for a '0.2.0' or
'1.0.0' version.